+++ example/android/app/src/main/kotlin/com/example/flutter_thermal_printer_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.flutter_thermal_printer_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}