+++ macos/Classes/FlutterThermalPrinterPlugin.swift
@@ -9,7 +9,7 @@ import CoreFoundation
 import AppKit
 import ApplicationServices
 
-public class FlutterThermalPrinterPlugin: NSObject, FlutterPlugin  , FlutterStreamHandler{
+public class BillyThermalPrinterPlugin: NSObject, FlutterPlugin  , FlutterStreamHandler{
     
     private var eventSink: FlutterEventSink?
     public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
@@ -23,10 +23,10 @@ public class FlutterThermalPrinterPlugin: NSObject, FlutterPlugin  , FlutterStre
     }
     
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let channel = FlutterMethodChannel(name: "flutter_thermal_printer", binaryMessenger: registrar.messenger)
-        let instance = FlutterThermalPrinterPlugin()
+        let channel = FlutterMethodChannel(name: "billy_thermal_printer", binaryMessenger: registrar.messenger)
+        let instance = BillyThermalPrinterPlugin()
         registrar.addMethodCallDelegate(instance, channel: channel)
-        let eventChannel =  FlutterEventChannel(name: "flutter_thermal_printer/events", binaryMessenger: registrar.messenger)
+        let eventChannel =  FlutterEventChannel(name: "billy_thermal_printer/events", binaryMessenger: registrar.messenger)
         eventChannel.setStreamHandler(instance)
     }