+++ test/unit/ble_config_test.dart
@@ -1,5 +1,5 @@
+import 'package:billy_thermal_printer/utils/ble_config.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_thermal_printer/utils/ble_config.dart';
 
 void main() {
   group('BleConfig', () {