+++ test/unit/printer_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_thermal_printer/utils/printer.dart';
+import 'package:billy_thermal_printer/utils/printer.dart';
 
 void main() {
   group('Printer', () {